+++ lib/biotcm/table.rb
@@ -1,7 +1,7 @@
 # encoding: UTF-8
 
 module BioTCM
-  # One of the basic data models used in BioTCM, 
+  # One of the basic data models used in BioTCM to process flat files, 
   # developed under <b>"strict entry and tolerant exit"</b> philosophy
   # (please refer to the test for details). 
   class BioTCM::Table