+++ HISTORY.md
@@ -1,6 +1,10 @@
 History
 ===============
 
+### By now 2014-03-15
+* 1 major enhancement
+	* Added Log module
+
 ### 0.0.1 2014-03-14
 * 2 major enhancement
 	* Completed the design of architectures